Failed Attempt to implement embedded PDFs
This is my attempt to allow embedding of PDF pages in other pages.
The good news is that it does not crash, the bad news is that it does
not work :)

Architecture-wise it was my idea was to follow the implementation of
how we embed PNGs:
Have an `embedPdfDocument` method  on PdfDocument which uses a
`PdfPageEmbedder`.
On PdfPage the returned object can be inserted using
`drawEmbeddedPdfPage`.

The PDF page to be embedded is wrapped into an XObject. This involved
some heavy wrestling against contexts and copyied pages.

I adapted the scratchpad to illustrate how this could be used. Sadly, it
just renders an empty, white page. I guess I need help debugging this.

Things to improve:

* I violated the PDFImage class by adding an EmbeddedPdfPage option to
  the existing image types. To me it sounds like a PDFPage is not an
image and should either not be supported by this class or we should
rename/rework this class
* XObject seem to support transformation matrices and transparency. This
  is great, but the normal PNG code scales/rotates their content
differently. We need to decide which is better
* Add tests, but maybe make it work before :)
* Reset the scratchpad to have the original code before merging this to
  master